requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
motor>=3.4
python-dotenv>=1.0
pydantic>=2.7.4